#1198 Add a abstract ScopingNode to encapsulate common scoping logic for Schedule and Container
